smart-money/requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
schedule>=1.2.0
